@@ -14,8 +14,9 @@ namespace igi {
         color3 _background;
 
       public:
-        META_B(scene, rflite::ctor([](const serializer_t &ser,
-                                      const std::pmr::polymorphic_allocator<void *> &alloc) {
+        using allocator_type = typename aggregate::allocator_type;
+
+        META_B(scene, rflite::ctor_a([](const serializer_t &ser, const allocator_type &alloc) {
                    assert(ser.HasMember("material"));
                    assert(ser.HasMember("surface"));
                    assert(ser.HasMember("entity"));
@@ -32,18 +33,19 @@ namespace igi {
                    color3 back = ser.HasMember("background")
                                      ? serialization::Deserialize<color3>(ser["background"])
                                      : palette::black;
-
-                   return scene(serialization::Deserialize<IMaterial *>(mser, alloc),
-                                serialization::Deserialize<IMaterial *>(sser, alloc),
-                                serialization::Deserialize<entity>(sser, alloc), alloc, back);
+                   
+                   std::pmr::vector<IMaterial *> mats = serialization::DeserializeArray<IMaterial *>(mser, alloc);
+                   std::pmr::vector<ISurface *> surfs = serialization::DeserializeArray<ISurface *>(sser, alloc);
+                   std::pmr::vector<entity> ents      = serialization::DeserializeArray<entity>(eser, alloc, mats, surfs);
+                   return scene(std::move(mats), std::move(surfs), std::move(ents), std::move(back), alloc);
                }))
         META_E
 
-        scene(std::pmr::vector<IMaterial *> &&mats, std::pmr::vector<ISurface *> &&surfs,
-              const std::pmr::vector<entity> &entities, typename aggregate::allocator_type &alloc,
-              color3 background = palette::black)
+        scene(std::pmr::vector<IMaterial *> mats, std::pmr::vector<ISurface *> surfs,
+              std::pmr::vector<entity> entities, const color3 &background,
+              const allocator_type &alloc)
             : _materials(std::move(mats)), _surfaces(std::move(surfs)),
-              _aggregate(entities, alloc), _background(background) { }
+              _aggregate(std::move(entities), alloc), _background(background) { }
 
         const aggregate &getAggregate() const {
             return _aggregate;
